Added option to clear the cache and use it by default
