+++ ugs-platform/ugs-platform-ugscore/src/main/java/com/willwinder/ugs/nbp/core/lifecycle/ServiceStarter.java
@@ -26,37 +26,18 @@ import com.willwinder.ugs.nbp.lib.services.LocalizingService;
 import com.willwinder.universalgcodesender.Utils;
 import com.willwinder.universalgcodesender.utils.Settings;
 import com.willwinder.universalgcodesender.utils.Version;
-import org.netbeans.api.sendopts.CommandException;
-import org.netbeans.spi.sendopts.Env;
-import org.netbeans.spi.sendopts.Option;
-import org.netbeans.spi.sendopts.OptionProcessor;
-import org.openide.cookies.OpenCookie;
-import org.openide.filesystems.FileUtil;
-import org.openide.loaders.DataObject;
-import org.openide.loaders.DataObjectNotFoundException;
 import org.openide.modules.OnStart;
 import org.openide.util.Lookup;
-import org.openide.util.lookup.ServiceProvider;
 import org.openide.windows.WindowManager;
 
-import java.io.File;
-import java.util.HashSet;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
-import java.util.logging.Level;
 import java.util.logging.Logger;
 
 /**
  * @author wwinder
  */
-@ServiceProvider(service = OptionProcessor.class)
 @OnStart
-public class startup extends OptionProcessor implements Runnable {
-    private static final Logger logger = Logger.getLogger(startup.class.getName());
-
-    private final Option openOption = Option.additionalArguments('o', "open");
-    private final Option defaultOpenOption = Option.defaultArguments();
+public class ServiceStarter implements Runnable {
+    private static final Logger logger = Logger.getLogger(ServiceStarter.class.getName());
 
     @Override
     public void run() {
@@ -90,52 +71,6 @@ public class startup extends OptionProcessor implements Runnable {
 
     private void setupVersionInformation(Settings settings) {
         // Only change the window title when all the UI components are fully loaded.
-        WindowManager.getDefault().invokeWhenUIReady(() -> {
-            Utils.checkNightlyBuild(settings);
-        });
-    }
-
-    /**
-     * Register interest in the "open" option.
-     */
-    @Override
-    public Set<Option> getOptions() {
-        HashSet<Option> set = new HashSet<>();
-        set.add(openOption);
-        set.add(defaultOpenOption);
-        return set;
-    }
-
-    /**
-     * CLI Handler.
-     */
-    @Override
-    protected void process(Env env, Map<Option, String[]> optionsMap) throws CommandException {
-        Optional<String> fileToOpen = getFileToOpen(optionsMap, openOption);
-        if (!fileToOpen.isPresent()) {
-            fileToOpen = getFileToOpen(optionsMap, defaultOpenOption);
-        }
-
-        if (fileToOpen.isPresent()) {
-            String inputFile = fileToOpen.get();
-            logger.info("File to open: " + inputFile);
-            try {
-                DataObject.find(FileUtil.toFileObject(new File(inputFile)))
-                        .getLookup()
-                        .lookup(OpenCookie.class)
-                        .open();
-            } catch (DataObjectNotFoundException e) {
-                logger.log(Level.SEVERE, "Could not open file " + inputFile, e);
-            }
-        }
-    }
-
-    private Optional<String> getFileToOpen(Map<Option, String[]> optionsMap, Option option) {
-        String[] files = optionsMap.getOrDefault(option, new String[0]);
-
-        if (files.length > 0) {
-            return Optional.of(files[0]);
-        }
-        return Optional.empty();
+        WindowManager.getDefault().invokeWhenUIReady(() -> Utils.checkNightlyBuild(settings));
     }
 }
